Refactor the form component to be more reusable
